Converted Mozilla Public License to new markup.
Dropped 'esp. as a subsection of your own opyright notice' because this was present
only in some of the use-cases and for no apparent reason.
Removed the requirement to inform the user about a download repository from the task
to establish that repository because this requirement was already captured by a separate task.

Fixed some typos and the Use-Case Names in the PHP-License.

Put the includes in oslic.tex each in its own group to make the commands defined
in the task lists local to their respective license.